requirements.txt
scikit-learn
joblib
flask
numpy
Pillow